requirements.txt
ddsp
matplotlib
numpy
soundfile
apache_beam
gin
gin-config
google-colab
absl-py==0.10
